Add wisp and totem

+++ Win32Project1/Include/ZeroProcessMonitor.h
@@ -12,7 +12,7 @@ public:
 	void Close();
 
 	float GetCpuUsage();
-	DWORD GetMemoryUsage();
+	size_t GetMemoryUsage();
 
 protected:
 	HRESULT CreateCpuUsageCounter();
